Terminates Clonor when exiting GUI
Added a simple edit to the script in which the process is terminated when you quit the GUI

Still having trouble trying to get the click button active again when the subprocess ends. Methods that do work end up freezing the GUI, which I don't want. Trying to understand how to terminate  a process while it is within a "thread", not sure if I should continue down this direction